Changes to get my front to back link to work

+++ src/main/java/com/revature/pokemondb/services/FanartService.java
@@ -63,8 +63,8 @@ public class FanartService {
 	 * @param id the id of the requested fanart object
 	 * @return the requested fanart object or null if the id does not exist
 	 */
-	public Fanart getFanart(int id) {
-		Optional<Fanart> artOpt = artRepo.findById(id);
+	public Fanart getFanart(long id) {
+		Optional<Fanart> artOpt = artRepo.findById((int) id);
 		if (artOpt.isPresent()) {
 			return artOpt.get();
 		} else {
